+++ example/pages/dialog/index.js
@@ -1,4 +1,5 @@
 const Dialog = require('../../dist/dialog/dialog');
+const Toast = require('../../dist/toast/toast');
 
 Page({
   toggleBaseDialog() {
@@ -73,6 +74,60 @@ Page({
       }]
     }).then(({ type }) => {
       console.log('=== dialog with vertical buttons ===', `type: ${type}`);
+    });
+  },
+
+  toggleOpenDialog() {
+    Dialog({
+      title: '弹窗',
+      message: '获取用户数据',
+      selector: '#zan-open-dialog',
+      buttons: [{
+        text: '用户信息',
+        type: 'userInfo',
+        openType: 'getUserInfo'
+      }, {
+        text: '获取手机号',
+        type: 'tel',
+        openType: 'getPhoneNumber'
+      }, {
+        text: '打开授权页',
+        type: 'setting',
+        openType: 'openSetting'
+      }]
+    }).then(({ type, hasOpenDataPromise, openDataPromise }) => {
+      console.log(type);
+
+      // 如果没有 open 数据返回，就不处理
+      if (!hasOpenDataPromise) {
+        return;
+      }
+
+      openDataPromise.then((data) => {
+        console.log('成功获取信息', data);
+      }).catch((data) => {
+        console.log('获取信息失败', data);
+      });
+    });
+  },
+
+  toggleCloseDialog() {
+    Dialog({
+      title: '弹窗',
+      message: '这是一个自由控制关闭的弹窗',
+      selector: '#zan-close-dialog',
+      autoClose: false
+    }).then(() => {
+      console.log('=== dialog resolve ===', 'type: confirm');
+      Toast.loading({
+        selector: '#zan-toast'
+      });
+
+      // 2s 后自动关闭弹窗
+      setTimeout(() => {
+        Dialog.close({ selector: '#zan-close-dialog' });
+        Toast.clear();
+      }, 2000);
     });
   },