+++ packages/card/index.js
@@ -3,7 +3,7 @@ Component({
     multipleSlots: true
   },
 
-  externalClasses: ['card-class'],
+  externalClasses: ['card-class', 'thumb-class'],
 
   properties: {
     useThumbSlot: {