+++ packages/field/index.js
@@ -1,6 +1,14 @@
 Component({
   behaviors: ['wx://form-field'],
 
+  externalClasses: ['field-class'],
+
+  relations: {
+    '../cell-group/index': {
+      type: 'parent'
+    }
+  },
+
   properties: {
     title: String,
     type: {
@@ -8,12 +16,12 @@ Component({
       value: 'input'
     },
     disabled: Boolean,
+    focus: Boolean,
     inputType: {
       type: String,
       value: 'text'
     },
     placeholder: String,
-    focus: Boolean,
     mode: {
       type: String,
       value: 'normal'
@@ -24,6 +32,10 @@ Component({
       type: Number,
       value: 140
     }
+  },
+
+  data: {
+    showBorder: true
   },
 
   methods: {
@@ -41,6 +53,17 @@ Component({
 
     handleFieldBlur(event) {
       this.triggerEvent('blur', event);
+    },
+
+    updateIsLastElement(isLastField) {
+      let showBorder = true;
+      if (isLastField && this.data.mode === 'normal') {
+        showBorder = false;
+      }
+
+      this.setData({
+        showBorder
+      });
     }
   }
 });