+++ packages/index.js
@@ -1,4 +1,4 @@
 exports.Dialog = require('./dialog/dialog');
 exports.Toast = require('./toast/toast');
-// exports.TopTips = require('./toptips/index');
+exports.TopTips = require('./toptips/toptips');