+++ packages/toast/toast.js
@@ -1,4 +1,5 @@
 const TOAST_CONFIG_KEY = 'zanui.__zanToastPageConfig';
+const DEFAULT_SHOW_TOAST_TIME = 3000;
 
 let timeoutData = {
   timeoutId: 0,
@@ -57,9 +58,16 @@ function Toast(optionsOrMsg, pageCtx) {
     show: true
   });
 
-  const timeoutId = setTimeout(() => {
-    toastCtx.clear();
-  }, parsedOptions.timeout || 3000);
+  let timeoutId = 0;
+  // toast 计时，如果小于0，就不会去关闭。
+  // 如果不传，就取默认值
+  const timeoutOption = parsedOptions.timeout || DEFAULT_SHOW_TOAST_TIME;
+  if (timeoutOption >= 0) {
+    timeoutId = setTimeout(() => {
+      toastCtx.clear();
+    }, timeoutOption);
+  }
+
 
   timeoutData = {
     timeoutId,
@@ -74,7 +82,7 @@ Toast.setDefaultOptions = function (options = {}, type = 'page') {
     type: options.type || '',
     icon: options.icon || '',
     image: options.image || '',
-    timeout: options.timeout || 3000
+    timeout: options.timeout || DEFAULT_SHOW_TOAST_TIME
   };
 
   if (type === 'global') {